Refactor config system: improve parsing, error handling, and module structure
