+++ CONFIG_FIXES.md
@@ -1,0 +1,172 @@
+# Critical Configuration System Fixes
+
+## Issue 1: Triple Parsing Performance Bug ❌
+
+### Problem
+The current implementation parses the SAME file content THREE times:
+```rust
+let contents = fs::read_to_string(filename)?;
+let simulation: SimulationConfig = toml::from_str(&contents)?;  // Parse 1
+let source: SourceConfig = toml::from_str(&contents)?;          // Parse 2  
+let output: OutputConfig = toml::from_str(&contents)?;          // Parse 3
+```
+
+This is:
+- **3x slower** than necessary
+- **Confusing** - repeated error messages for syntax errors
+- **Non-idiomatic** - violates standard serde patterns
+
+### Solution
+Parse ONCE into a helper struct:
+
+```rust
+// Helper struct matching TOML file structure
+#[derive(Deserialize)]
+struct TomlConfigHelper {
+    simulation: SimulationConfig,
+    source: SourceConfig,
+    output: OutputConfig,
+}
+
+impl Config {
+    pub fn from_file(filename: &str) -> Result<Self, ConfigError> {
+        debug!("Loading config from {}", filename);
+        let contents = fs::read_to_string(filename)?;
+        
+        // Parse ONCE
+        let helper: TomlConfigHelper = toml::from_str(&contents)?;
+        
+        Ok(Self {
+            simulation: helper.simulation,
+            source: helper.source,
+            output: helper.output,
+        })
+    }
+}
+```
+
+## Issue 2: Missing Save Functionality ❌
+
+### Problem
+`FieldValidationConfig` has both load/save, but main `Config` can only load.
+Users cannot:
+- Save modified configurations
+- Export default configs
+- Ensure reproducibility
+
+### Solution
+Add `to_file` method with proper serialization:
+
+```rust
+// Add Serialize derive
+#[derive(Debug, Serialize, Deserialize)]
+pub struct Config {
+    pub simulation: SimulationConfig,
+    pub source: SourceConfig,
+    pub output: OutputConfig,
+}
+
+// Helper for serialization (references avoid cloning)
+#[derive(Serialize)]
+struct TomlSerializeHelper<'a> {
+    simulation: &'a SimulationConfig,
+    source: &'a SourceConfig,
+    output: &'a OutputConfig,
+}
+
+impl Config {
+    pub fn to_file(&self, filename: &str) -> Result<(), ConfigError> {
+        let helper = TomlSerializeHelper {
+            simulation: &self.simulation,
+            source: &self.source,
+            output: &self.output,
+        };
+        
+        let contents = toml::to_string_pretty(&helper)?;
+        fs::write(filename, contents)?;
+        Ok(())
+    }
+}
+```
+
+## Issue 3: String Error Anti-Pattern ❌
+
+### Problem
+```rust
+pub fn from_file(filename: &str) -> Result<Self, String>  // BAD!
+```
+
+Using `String` for errors:
+- Loses error type information
+- Prevents programmatic error handling
+- Inconsistent with rest of codebase using `thiserror`
+
+### Solution
+Use proper error enum with `thiserror`:
+
+```rust
+// Enhance existing ConfigError
+use thiserror::Error;
+
+#[derive(Error, Debug)]
+pub enum ConfigError {
+    #[error("Failed to read configuration file: {0}")]
+    Io(#[from] std::io::Error),
+    
+    #[error("Failed to parse TOML: {0}")]
+    TomlParse(#[from] toml::de::Error),
+    
+    #[error("Failed to serialize TOML: {0}")]
+    TomlSerialize(#[from] toml::ser::Error),
+    
+    // ... existing variants
+}
+
+// Now errors propagate automatically with ?
+impl Config {
+    pub fn from_file(filename: &str) -> Result<Self, ConfigError> {
+        let contents = fs::read_to_string(filename)?;  // Io error
+        let helper: TomlConfigHelper = toml::from_str(&contents)?;  // TomlParse error
+        Ok(Self { /* ... */ })
+    }
+}
+```
+
+## Issue 4: Misplaced FieldValidationConfig ❌
+
+### Problem
+`FieldValidationConfig` is in main `config/mod.rs` but it's a validation concern.
+Meanwhile, there's a `config/validation.rs` module that would be its natural home.
+
+### Solution
+Move to proper module:
+
+```rust
+// Move from src/config/mod.rs to src/config/validation.rs
+pub struct FieldValidationConfig { /* ... */ }
+pub struct FieldLimits { /* ... */ }
+
+// In src/config/mod.rs, just re-export:
+pub use validation::{FieldValidationConfig, FieldLimits};
+```
+
+## Performance Impact
+
+- **Parsing**: 3x faster (parse once vs thrice)
+- **Error handling**: Zero-cost with proper From impls
+- **Memory**: Negligible (helper structs are temporary)
+
+## Correctness Impact
+
+- **Error handling**: Structured errors enable proper recovery
+- **Reproducibility**: Save functionality ensures exact config preservation
+- **Organization**: Logical module structure improves maintainability
+
+## Implementation Priority
+
+1. **IMMEDIATE**: Fix triple parsing (performance bug)
+2. **HIGH**: Fix error types (API consistency)
+3. **MEDIUM**: Add save functionality (feature completeness)
+4. **LOW**: Move validation config (code organization)
+
+These aren't stylistic preferences - they're fundamental corrections to a broken configuration system.
